resolved creating, editing, deleting post photos with better strategy

+++ api/src/api/post/post.services.ts
@@ -62,34 +62,24 @@ const updatePost = async (
   title: string,
   rating: number,
   content: string,
-  postPhotos: UploadedPhoto[]
+  postPhotos: UploadedPhoto[],
+  deletePhotos: UploadedPhoto[]
 ) => {
-  // for each photo use prisma client and delete them, then recreate them, easier than
-  // doing multiple put requests, we can optimize later. Just delete all, and then create all.
-  // front-end needs to change for this as well, no filtering is required.
+  if (deletePhotos[0]) {
+    deletePhotos.forEach(async (photo) => {
+      if (photo.id) {
+        await deletePostPhoto(photo.id);
+      }
+    });
+  }
 
-  // Get all the images
-  const currentPost = await getPost(postId);
-
-  // Deleting
-  await currentPost?.postPhotos.forEach(async (photo) => {
+  postPhotos.forEach((photo) => {
     if (photo.id) {
-      await deletePostPhoto(photo.id);
+      updatePostPhoto(photo.url, photo.id);
+    } else {
+      createPostPhoto(photo.url, postId);
     }
   });
-
-  // Add all of the photos back to the database
-  await postPhotos.forEach((photo) => {
-    createPostPhoto(photo.url, postId);
-  });
-
-  // postPhotos.forEach((photo) => {
-  //   if (photo.id) {
-  //     updatePostPhoto(photo.url, photo.id);
-  //   } else {
-  //     createPostPhoto(photo.url, postId);
-  //   }
-  // });
   const post = await prismaClient.post.update({
     where: {
       id: postId,
